backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "backend"
version = "1.0.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "cryptography>=44.0.2",
    "fastapi[standard]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "sentry-sdk>=2.25.1",
    "sqlalchemy[mypy]>=2.0.40",
    "tenacity>=9.1.2",
    "types-passlib>=1.7.7.20250408",
    "uvicorn>=0.34.0",
    "aioredis>=2.0.1",
    "orjson>=3.10.16",
    "slowapi>=0.1.9",
    "faststream[cli,rabbit]>=0.5.39",
    "aiosmtplib>=4.0.0",
    "fastapi-profiler>=1.4.1",
    "aiosqlite>=0.21.0",
    "sqlalchemy-utils>=0.41.2",
    "asgi-lifespan>=2.1.0",
    "email-validator>=2.2.0",
    "pydantic>=2.11.3",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "memory-profiler>=0.61.0",
    "psutil>=7.0.0",
    "pandas>=2.3.0",
    "types-psutil>=7.0.0.20250601",
    "pandas-stubs>=2.2.3.250527",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "tabulate>=0.9.0",
    "pre-commit>=4.2.0",
    "types-psycopg2>=2.9.21.20250718",
]

[dependency-groups]
dev = [
    "bandit>=1.8.5",
    "fakeredis>=2.30.1",
    "freezegun>=1.5.1",
    "httpx>=0.28.1",
    "locust>=2.37.11",
    "memory-profiler>=0.61.0",
    "mypy>=1.15.0",
    "psutil>=7.0.0",
    "pyinstrument>=5.0.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-benchmark>=5.1.0",
    "pytest-cov>=6.2.1",
    "pytest-lazy-fixtures>=1.3.2",
    "pytest-mock>=3.14.1",
    "pytest-sqlalchemy-mock>=0.1.7",
    "pytest-xdist>=3.8.0",
    "ruff>=0.11.5",
    "vulture>=2.14",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
cache-dir = ".cache/ruff"
exclude = [".venv", ".cache", ".benchmarks", "alembic"]

[tool.ruff.format]
docstring-code-format = true
line-ending = "lf"

[tool.mypy]
python_version = "3.13"
cache_dir = ".cache/mypy"
strict = true
show_error_codes = true
exclude = [".venv", ".cache", ".benchmarks", "alembic", "scripts", "monitoring", "tests/parametrized_test_data.py"]
disallow_untyped_defs = true
disallow_any_unimported = false
ignore_missing_imports = false
follow_imports = "normal"
explicit_package_bases = true
no_implicit_optional = false

plugins = [
    "pydantic.mypy",
    "sqlalchemy.ext.mypy.plugin"
]

[[tool.mypy.overrides]]
module = "gunicorn.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.parametrized_test_data"
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
cache_dir = ".cache/pytest"
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "--strict-markers",
    "--strict-config", 
    "--verbose",
    "--tb=short",
    "--maxfail=10",
    "--benchmark-storage=./.benchmarks"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "performance: Performance tests", 
    "contract: Contract tests",
    "slow: Slow running tests"
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::pytest.PytestUnhandledCoroutineWarning",
    "ignore:coroutine.*was never awaited:RuntimeWarning"
]

[tool.coverage.run]
branch = true
source = ["app"]
omit = [
    "tests/*",
    "scripts/*",
    "app/core/config/*",
    "app/core/exceptions/*",
    "app/core/profiler.py",
    "app/backend_pre_start.py",
    "app/main.py"
]

[tool.coverage.report]
omit = [
    "tests/*",
    "scripts/*",
    "app/core/config/*",
    "app/core/exceptions/*",
    "app/core/profiler.py",
    "app/backend_pre_start.py",
    "app/main.py"
]

[tool.coverage.html]
directory = "htmlcov"

[tool.bandit]
exclude_dirs = [".venv", ".cache", ".benchmarks", "alembic", "scripts", "monitoring", "tests", "app/utils/cache.py"]

[tool.vulture]
exclude = [".venv", ".cache", ".benchmark"]
ignore_decorators = ["@limit", "@cached", "@invalidate_cache"]
min_confidence = 80
sort_by_size = true
verbose = true
